refactor: include section title and description in section layout component itself
