--- src/components/shared/Section.jsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import { Heading1, Subtitle, SectionContainer } from '.';
-
-function Section({ children, title, description }) {
-  return (
-    <SectionContainer>
-      <Heading1>{title}</Heading1>
-      <Subtitle>{description}</Subtitle>
-      {children}
-    </SectionContainer>
-  );
-}
-
-export default Section;
