--- src/components/shared/Typography/Subtitle.jsx
@@ -1,16 +1,0 @@
-import styled from 'styled-components';
-import tw from 'twin.macro';
-
-export default styled.p`
-  ${tw`
-        font-normal
-        text-base
-        2xl:text-base
-        lg:text-base
-        md:text-base
-        sm:text-sm
-        text-color-secondary
-        font-roboto
-        w-100
-    `}
-`;
